- Casino class has been implemented

+++ src/main/java/duckCasino/behaviors/fly/NoFlyingDuckBehaviour.java
@@ -2,7 +2,7 @@ package duckCasino.behaviors.fly;
 
 public final class NoFlyingDuckBehaviour implements FlyBehavior {
     @Override
-    public double fly(int time) {
+    public int fly(int time) {
         return 0;
     }
 }
